# Cache __toString() representation