Some libraries perform many implicit calls to `Uri::__toString()`, which heavily affect their performance.

For example, [in spatie/crawler](https://github.com/spatie/crawler/blob/67974a810f6a342c333c4e348fcfad7c1375ffde/src/CrawlQueue/CollectionCrawlQueue.php#L103-L117):

```php
protected function contains($collection, CrawlUrl $searchCrawlUrl): bool
{
    foreach ($collection as $crawlUrl) {
        if ((string) $crawlUrl->url === (string) $searchCrawlUrl->url) {
            return true;
        }
   }
   return false;
}
```

Of course, this specific example could be improved by moving the rightmost `(string)` conversion out of the loop; but the leftmost conversion is harder to optimize: every call to `contains()` will perform a `(string)` conversion on every element in the collection, making it harder to optimize without refactoring the whole class.

I was able to speed up the above library by an order of magnitude with a simple fix: **cache the `__toString()` output**.

This is perfectly safe, as PSR-7 requires that `Uri` objects are immutable.

I guess this improvement can speed-up a lot of other libraries, with very little effort. I hope this change can go upstream 👍