@Tobion 2019-08-06
The logic of `composeComponents` is very little that it does not seem worth to me. At the end this is just trading memory for CPU. So I'm hestitant to merge this as it can be solved in user-land.