@BenMorel 2019-08-06
> The logic of composeComponents is very little that it does not seem worth to me. At the end this is just trading memory for CPU.

It's little, but enough to make a big difference in my benchmarks. I do understand your concerns, though.