@BenMorel 2019-08-06
Micro-benchmarks:

**CPU**

```php
$t = microtime(true);

for ($i = 0; $i < 1000000; $i++) {
        $str = (string) $uri;
}

printf("%.3f s.\n", microtime(true) - $t);
```

**Memory**

```php
$uris = [];

$base = memory_get_usage();

for ($i = 0; $i < 10000; $i++) {
        $uris[] = $uri = new Uri('https://www.example.com/foo/bar/baz?a=123&b=456&c=789#fragment');
        $str = (string) $uri;
}

printf("%.3f KiB\n", (memory_get_usage() - $base) / 1024);
```
**Results**

|        | Before   | After    | Diff  |
|--------|----------|----------|-------|
| CPU    | 2179 ms  | 525 ms   | 4.15x |
| Memory | 4246 KiB | 5497 KiB | +29%  |

I guess this is a tradeoff only you can decide upon!