@BenMorel 2019-08-06
A last note:

> So I'm hestitant to merge this as it can be solved in user-land.

This can for sure be implemented in userland, although it complicates the implementation quite a bit. Take this [CrawlUrl](https://github.com/spatie/crawler/blob/master/src/CrawlUrl.php) class for example, we would have to change it from:

```php
    /** @var \Psr\Http\Message\UriInterface */
    public $url;

    /** @var \Psr\Http\Message\UriInterface */
    public $foundOnUrl;
```

to:

```php
    /** @var \Psr\Http\Message\UriInterface */
    public $url;

    /** @var string */
    public $urlString;

    /** @var \Psr\Http\Message\UriInterface */
    public $foundOnUrl;

    /** @var string */
    public $foundOnUrlString;
```

Which is, to say the least, cumbersome. Having the ability to pass a `Uri` around, to use it either as an object or as a string, without caring about performance is a big plus IMHO.