@Tobion 2019-08-07
I'm accepting this as the additional memory will get released when the uri is not used anymore. So it shoudn't hurt anyone.