@Tobion 2019-08-07
See https://github.com/guzzle/psr7/releases/tag/1.6.0