@BenMorel 2019-08-07
> which will likely be the last minor release in 1.x

Last minor, but this one could go in a patch release?